# Conflicts:
#	user-interface/src/main/bundles/dev.bundle
#	user-interface/src/main/java/life/qbic/datamanager/views/projects/project/measurements/MeasurementMetadataUploadDialog.java

+++ project-management/src/main/java/life/qbic/projectmanagement/application/measurement/MeasurementService.java
@@ -554,7 +554,7 @@ public class MeasurementService {
     if (value.isBlank()) {
       return -1;
     }
-    return Integer.parseInt(value);
+    return (int) Double.parseDouble(value);
   }
 
   /**
